refactor: extract useEstimateForm hook from NewEstimateForm (Phase 4)
Moved all state (28 useState), effects, handlers, and computed values
from NewEstimateForm.tsx (2484 lines) into hooks/useEstimateForm.ts.
Component is now purely JSX, calling the hook for all state and behavior.
